+++ src/main/java/data/hirwan.txt
@@ -1,3 +1,5 @@
-. [T][X] faj
+. [T][ ] faj
 . [T][X] fja
 . [T][ ] kasfj
+. [T][ ] fajfj
+. [T][ ] jfdaj
