+++ src/main/java/hirwan/Ui.java
@@ -1,0 +1,22 @@
+package hirwan;
+
+import java.io.*;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Scanner;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+
+class Ui {
+    public static String input() {
+        Scanner scanner = new Scanner(System.in);
+        String text = scanner.nextLine();
+        return text;
+    }
+
+    public static void output(String printText) {
+        System.out.println(printText);
+    }
+}
